+++ lib/src/brands/linkedin.dart
@@ -1,0 +1,16 @@
+import '../palette_rgb.dart';
+
+/// Linkedin's official brand colors:
+///
+/// - Blue;
+/// - White.
+///
+/// See also:
+/// - [linkedin colors](https://brandpalettes.com/linkedin-color-codes/)
+class Linkedin extends PaletteRGB {
+  /// Blue #2867B2.
+  const Linkedin.blue() : super.opaque(0x2867B2);
+
+  /// White #FFFFFF.
+  const Linkedin.white() : super.opaque(0xFFFFFF);
+}